+++ semana-14/aula45/exercicio01/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "aula45",
+  "name": "exercicio01",
   "version": "1.0.0",
-  "description": "node e package",
+  "description": "",
   "main": "index.js",
   "scripts": {
     "start": "tsc && node ./build/index.js",
@@ -11,7 +11,6 @@
   "license": "ISC",
   "dependencies": {
     "@types/node": "^14.0.13",
-    "ts-node": "^8.10.2",
     "typescript": "^3.9.5"
   }
 }
