+++ semana-14/aula45/exercicio01/src/index.ts
@@ -1,0 +1,9 @@
+//EXERCICIO 01
+
+const nome: string = process.argv[2];
+const idade = Number(process.argv[3]);
+
+function programa(nome: string, idade: number){
+    console.log(`Olá, ${nome}! Você tem ${idade} anos. Em sete anos você terá ${idade+7}.`)
+}
+programa(nome, idade)
