+++ semana-14/aula45/exercicio01/tsconfig.json
