+++ semana-14/aula45/exercicio03/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "exercicio03",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "start": "tsc && node ./build/index.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Nathalia_Costa",
+  "license": "ISC",
+  "dependencies": {
+    "@types/node": "^14.0.13",
+    "typescript": "^3.9.5"
+  }
+}
